--- FreeCam/MapControllerPatches.cs
@@ -1,19 +1,0 @@
-﻿using HarmonyLib;
-
-namespace FreeCam;
-
-[HarmonyPatch(typeof(MapController))]
-internal static class MapControllerPatches
-{
-	// Fixes bug where using the freecam in map mode keeps you stuck in the None input mode
-
-	private static InputMode _oldInputMode;
-
-	[HarmonyPrefix]
-	[HarmonyPatch(nameof(MapController.EnterMapView))]
-	public static void MapController_EnterMapView() => _oldInputMode = OWInput.GetInputMode();
-
-	[HarmonyPostfix]
-	[HarmonyPatch(nameof(MapController.ExitMapView))]
-	public static void MapController_ExitMapView(MapController __instance) => OWInput.ChangeInputMode(_oldInputMode);
-}
